@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
 import App from './routes/App';
@@ -9,9 +9,10 @@ import initialState from './initialState';
 
 const store = createStore(reducer, initialState);
 
-ReactDOM.render(
+const container = document.getElementById('app');
+const root = ReactDOM.createRoot(container);
+root.render(
   <Provider store={store}>
     <App />
-  </Provider>,
-  document.getElementById('app'),
+  </Provider>
 );
