@@ -1,5 +1,5 @@
 import React from 'react';
-import { BrowserRouter, Switch, Route } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Home from '../containers/Home';
 import Checkout from '../containers/Checkout';
 import Layout from '../components/Layout';
@@ -8,11 +8,11 @@ import NotFound from '../containers/NotFound';
 const App = () => (
   <BrowserRouter>
     <Layout>
-      <Switch>
-        <Route exact path="/" component={Home} />
-        <Route exact path="/checkout" component={Checkout} />
-        <Route component={NotFound} />
-      </Switch>
+      <Routes>
+        <Route exact path="/" element={<Home />} />
+        <Route exact path="/checkout" element={<Checkout />} />
+        <Route element={<NotFound />} />
+      </Routes>
     </Layout>
   </BrowserRouter>
 );
